requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-dotenv
azure-storage-blob
openai
faiss-cpu
PyMuPDF
python-docx
tiktoken
azure-search-documents==11.6.0b12
azure-search-documents 
azure-identity
langchain 
langchain-openai
azure-ai-openai 
azure-core
langchain-community
loguru         